implemented the service classes, added a timer, going to start transfering this all over into gui and the controller soon. fingers crossed it all goes well. also changed a lot of things from ints to doubles to make my life easier.

+++ src/main/java/seng201/team124/models/Difficulty.java
@@ -53,7 +53,7 @@ public enum Difficulty {
      * @param baseMoney Base money amount
      * @return starting money adjusted for difficulty
      */
-    public int calculateStartingMoney(int baseMoney) {
+    public double calculateStartingMoney(double baseMoney) {
         return (int) (baseMoney * getMoneyMultiplier());
     }
 }
